remove improper call to CAMLparam from custom_operations finalization routine
According to the "Interfacing C with OCAML" documentation, the "finalize" routine
of a block allocated with alloc_custom must not trigger a garbage collection, and
therefore should not use the CAMLparam or CAMLreturn macros. The code was
incorrectly calling CAMLparam, which triggered a segfault when performing a major
collection of the garbage collector.