Add support for Events. #59
The `event` statement / construct is similar to `send`, with
the difference being that it has a name and takes only one `Message`
as argument. The implementation piggybacks on the `Msg` literal and
`Message` expr constructs in Syntax.ml.

An example usage of events is now added in the crowdfunding contract.
In particular, crowdfunding/output_1.json shows how events appear in
the output of the interpreter.

The typechecker has been extended to ensure that all literals used
in constructing a `Msg` (used in both `send` and `event` statements)
are storable.