Add a new ByStrX type, where X is a positive integer.
This type can represent arbitrary binary data (as hexadecimal
strings). With this, the Address and Hash types are now removed
(as they can be represented by ByStr20 and ByStr32 respectivel).

This change was done inorder to support more binary data types
for Schnorr signature feature, which requires ByStr32, ByStr33 and
ByStr64, all of which (and any arbitrary size) are now available.

GitHub Issue #188.