143 multistage level 2 (#194)
This PR contains the second part of the refactoring that will eventually allow user-defined checks to be plugged into the compiler (Issue #143). It is not the complete solution, but most future merge conflicts related to this issue can be avoided if the PR is merged now.

Datatypes for statements and expressions have been moved inside the ScillaSyntax functor, and the typechecker functions for statements and expression have similarly been moved inside the typechecker functor.

The changes to the typechecker makes it possible to extract the pattern match checker and have it work as a separate phase.

Error functions that previously resided in the typechecker have been moved to the Syntax functor, because they pattern match on syntax constructors. This move allows for phases to define their own minimal error functions that dispatch the creation of the actual error message to the Syntax functor, which is more modular and creates more systematic error messages. A new issue will be created shortly to parameterize the error functions with the name of the failing phase.

Finally, the type_checker and scilla_checker runners have been minimally modified to run the pattern match checker on the typechecked expression/module. The runners still need to be modified to use the various functors in a more systematic way.