Rich literals (#260)
An extensive refactoring, extending literals with two new kinds: for closures and type abstractions. Those literals replace previously used run-time values `ValClosure` and `ValTypeFunction`, allowing for more uniform treatment of runtime values. It also allowed to get rid of additional run-time abstractions, so Eval.ml now operates exclusively with literals.

The small downside of this unification is that some additional checks need to be in place, to make sure that instances of literals `Clo` and `TAbs` wouldn't be attempted to serialize or store to fields. for instance, the corresponding check needs to be inserted into Conf.store.

As the result the data type of runtime values `EvalUtil.value` is now gone altogether, as all values are accurately represented by literals.