rename scilla.el to scilla-mode.el (#404)
* rename scilla.el to scilla-mode.el

In order to enable compatibility with emacs' named features[0] (which
is very commonly used, especially by the use-package ecosystem), the
provided feature needs to match the filename which provides it.
However the 'scilla-mode symbol is currently provided by scilla.el,
not scilla-mode.el.  This results in an error like:

   Lisp error: (error "Loading file /path/to/scilla/misc/emacs-mode/scilla.el failed to provide feature ‘scilla’")

So rename the file to match the feature symbol.  This allows it to
be loaded via

    (use-package scilla
      :require flycheck
      :load-path (lambda () (concat scilla-root "/misc/emacs-mode")))

or similar, and brings it in line with many other major modes, e.g.

    $ find /usr/share/emacs | egrep -c -- '-mode\.el(\.gz)?$'
    56

[0] https://www.gnu.org/software/emacs/manual/html_node/elisp/Named-Features.html

* Update scilla-mode.el

* Update scilla-mode.el