Make release builds by default.
This prevents warnings from being flagged as errors.

To still pay attention to warnings, add a "dev" Makefile target
that does a developer build, which errors on warnings.

Travis uses "make test", which is based on "make dev" and hence
will trigger errors on warnings.
Travis also uses "make coveralls" which is based on "make" (release)
and hence will test that also.