Add support for scanning init.json for libraries.
This allows us to look at name -> address mapping of libraries
and look for "address.scillib" (and "address.json") files, which,
according to new convention will be the filename used by the blockchain.
This new convention allows for same named distinct libraries (which are
guaranteed to have different addresses).

During this, I've also fixed a bug in Testcontracts.ml that didn't
compare outputs of "expected to fail" tests.