Gas usage analysis for Scilla (#519)
Introducing a static resource analyzer for Scilla.

The analysis computes gas usage of a transition as a polynomial of the size of its parameters and contract fields. The gas usage of a higher-order functions may depend on the gas
usage of its functional argument, as well as on the size of the result value it returns. We capture
this by means of `sizeref` and `guref` (for gas usage reference) that together form a signature. The signature, allowing for higher-orderness can take arguments for substitution later.

This is still a work in progress and some of the TODOs are listed in #519. (There will be separate Issues created and kept track of after this merge).

Commit history:

* multivariate polynomial utility

* add expression analyser. not all expressions supported

* When substituting formals with actuals, resolve actuals from environment.

* Add signature printing functions

* Fix bug in creating lambda. Resolve arguments first.

* resolve inputs to builtins and fix result sizeref.

* support parameter expansion for polynomials

* Resolve and expand sizeref and guref lambdas

* substitution of actuals should replace lambda function names also.

Only a special case where the function being called is direct
(as against wrapped in ADTs) is supported.

* Make GApp printing more friendly

* gas-analysis: Support builtins

* add statement, transition and contract analysis

* add a testsuite for gas analysis

* Substitute all arguments at once, not one by one.

This fixes bug in analyzing tests/gas_use_analysis/expr/app2.scilla.
Previously, substitutions such as [a;b] -> [b;a] in [a;b] would result
in [a;a] instead of [b;a].

* support arbitrary functions of sizerefs

* simplify size references

* Replace BranchTaken sizeref with MaxB. Use for result of MatchExpr.

2. The option to not flatten (add up) gas-use across branches is removed.
3. Simplify Pair sizerefs in result.

* rename Component to Element

* add files missing in previous merge conflict resolution

* use sizeref polynomials to track sizes

* update test based on latest merge

* don't max over pattern matches, just add them up

* remove baseref as it isn't needed now

* Improve result size of the Cons constructor

* add a max_combine function for polynomials

* fix bug in max_combine

* use max_combine_pn instead of add_pn to combine match branches

* add a miniature (hacky) recurrence solver

* modernize uses of PrimTypes based on the new representations for it

* Update gas-analysis to use tree structured external libraries.

* Determine as Intractable when cannot determine recurrence

* Isolate polynomial utility

* Create an interface for Polynomial.ml