Decouple state accesses from Eval
Setup an outline for IPC based state accesses that can
fetch and update state variables on demand, each time fetching
and accessing only those parts of the state that is required.
The IPC part is, of course, not implemented yet. However this this
change, implementing IPC state access should no more involve modifications
to Eval or any part of Scilla except for the newly added StateService module.

Fixes #569