A few generalizations to the recurrence solver.
Note that with these changes, we are doing slightly worse
in reporting the result size of `list_tail.scilexp` which
previously had "equal to input list size", but now is
"equal to input list size + 1", while in reality `list_tail`
actually reduces the size by 1.

This is arising because of adding a `Container` and a simple
constant on two branches of the pattern match in `list_tail`.

This also affects, in turn, `list_eq` which uses `list_tail`.