Do not incentivize stores that reduce storage.
Incentivising stores that reduce the on-disk storage aren't
really useful because they are only a discount on the current
transaction and not pay back. For every transaction involving
a large stored data object, computation proportional to the
size is still charged. On the long run, it is better that we
move to a different model that charges correctly for persistant
storage.

On a side note: The current model does not work with the new
IPC based state accesses, since the previous size of a data
being stored is not available at all (to compute old_size - new_size)
to decide the incentive.