Efficient state accesses via IPC (#644)
* Basic skeleton

* Add protobuff files + client changes

* Add functions to (de)serialize protobuf values.

* Add ocaml-protoc and ppx_deriving_rpc dependencies to scilla.opam

* Do not output full state during deploying if is_ipc

* Fix bug in parsing JSON values.

* Handle exceptions thrown during JSON parsing and socket connection.

* Rename protobuf file so that auto-generated files don't need renaming

* Print field_depth_map in interpreter's deployment output

* Revert "Print field_depth_map in interpreter's deployment output"

* Fix minor bug in EvalUtil.ml : map_get

* scilla ipc tests (#623)

* Add basic server code

* Add json parsing code

* Add code for fetching

* Rename protobuff

* Add 'implementation' to IDL

* Implement update and fix most build errors

* Replace fail with jsonrpc error

* Replace code and messages

* Refactor + add license

* Integrate the mock server into the testuite, commented out.

* StateIPCTest: Do not update _balance to server

* Get the test IPC server (mostly) working.

The IPC tests themselves are still disabled in Testcontracts.ml.
There are two things to be done before these are enabled:
   a. Since the existing state.json is used to initialize the database
      of the mock server, it needs to be parsed. However, the state JSONs
      of contracts that use custom ADTs cannot be parsed without knowing
      the types (from the contract definition). It looks like I may have
      to first run the checker, note the custom ADTs and then parse the JSONs.
      The alternate is to make the parsing type agnostic (which is feasible
      since the IPC code doesn't care about types other than maps).
      This will however prevent a lot of code reuse (in particular,
      reusing the IPC client we have in the main code base for the testsuite).
   b. There are some issues with threading. The testsuite runs multiple
      threads / processes, but currently we can only have one state
      server running globally (entire system). I need to find a
      solution for this too.

* Use Caml.Hashtbl in StateIPCTestServer for same order of key-val with Scilla

* StateIPCTest: Use a separate IPC client to initialize the server

rather than using StateIPCClient. This is necessary since StateIPCClient
is based on Scilla literals, which cannot always be parsed from
state.json (use of custom ADTs make literals un-parsable without
the ADT definitions being parsed from the source).

* Fix minor bug in EvalUtil.ml : map_get

* testsuite: input state json must contain all fields

* Sort output of StateIPCTestServer to match that of gold, to avoid
spurious diffs when comparing with gold output.

Also an attempt at multi-threading StateIPCTestServer. It isn't
complete yet and the server doesn't yet shutdown effectively.

* Couple bug fixes

* Change server module to use persistent set of threads (thread pool).

Tests have been enabled but they still fail. There is an issue with
the file descriptors not being correctly closed and hence they build
up eventually running out.

* StateIPCClient: Use raw Unix API instead of wrappers.

* Fix bug in StateIPCTestClient. Phew.

We need to use Unix sockets directly and *close them*. Using channels
and just calling `Unix.shutdown_connection` doesn't work.

Diffing of output is currently disabled due to gas cost differences.
(see `StateIPCTest.ml:sort_mapkeys`)

* Make IPC tests multi-threaded, since the testsuite itself is

* Do not incentivize stores that reduce storage.

Incentivising stores that reduce the on-disk storage aren't
really useful because they are only a discount on the current
transaction and not pay back. For every transaction involving
a large stored data object, computation proportional to the
size is still charged. On the long run, it is better that we
move to a different model that charges correctly for persistant
storage.

On a side note: The current model does not work with the new
IPC based state accesses, since the previous size of a data
being stored is not available at all (to compute old_size - new_size)
to decide the incentive.

* Update costs for storing fields and remove workaround in JSON comparison for gas diffs

* Refactor UsefulLiterals to be more useful

* Simplify namespace handling

* Regroup related functions in UsefulLiterals

* Refactor map_get using UsefulLiterals

* Add a contract to test IPC map access corner cases.

Also add a script to run all the transitions sequentially
assuming the presence of an IPC state server.

* IPC map corner tests: Add a comment referencing the script to run

* Add map_corners_test to the testsuite

* Provide Makefile and testsuite options for testing against external IPC server

* Make contract_tests the first in the testsuite for fixed test_id

* In map_corners_test.scilla, ensure deleted sub-map is empty.

More fixes based on review comments.

* One more sub map deletion check.

* Handle exceptions encoding/decoding protobuf

* Minor variation to map_corners_test.scilla:t7

* improve comment in t7

* Reword comments around StateService fetch and update