Track precision and conditionals for exprs
The linearity & ops analysis is necessary but not sufficient to
determine which writes are commutative. The problem is that match
expressions introduce imprecision in the contribution type given
to an expression IF the clauses have different contributions.

To infer commutativity, we need to keep track of whether
contribution types are precise. (We can make positive decisions
only based on precise types. Negative decisions can rely on
imprecise types.)

We also keep track of which contributions flow into conditionals.
These contributions must be owned by the processing shard.