External libs, improved proc calls, effect dependencies
Added support for external libraries.

Improved the logic for procedure calls: a procedure call does not
need ALL its arguments to be parameters of the callee; it suffices
that the arguments that are used as keys in the called procedure
are parameters of the callee.

Now tracking dependencies of effects, including events, messages,
and control flow (conditionals). We still need to somehow account
for implicit effects, e.g. potential integer overflows.