Support closures
This commit gives accurate types to closures. Key to this is
introducing de Bruijn levels (similar to indices, but counting
from the outside-in rather than inside-out) in the `expr_type`
and `FormalParameter`. As such, the type of a function that is
a closure can refer to bindings defined in the closed context.

Example:

```
(* closure_test: {_0, Linear, add} {_0, Linear, add}
let closure_test = fun (a : Uint128) => fun (b : Uint128) =>
  (* f: {1._1, Linear, add}, {_1, Linear, add} *)
  let f = fun (a : Uint128) => fun (x : Uint128) => builtin add x b in
  f b a
```

The type of `f` refers to `_1`, i.e. `0._1`, which is a binding
that is closed over.

Application of functions takes into account the level of the function
(i.e. which level its formal parameters belong to). Note that a function's
level is NOT always the highest level in its type: a function might not use
its arguments, i.e. only use bindings in the closure. We store a function's
level in `expr_type`.

This introduces a slight problem, where the same function might have
different types, depending on which level it is given. We might need
some sort of normalisation, but I haven't thought too much about that
for now.