Detect read after write and mark as AlwaysExclusive
This maintains the invariant that all Pseudofields referred to
in expr_type are "fresh", i.e. the value of the field at the
beginning of the contract execution.

IMPORTANT: this is true only as long as the transition parameters
used as keys have different values. E.g. Transfer(from=aaa, to=aaa)
will alias and the blockchain code must detect this situation and
discount the summary/analysis if that is the case.