Generate global sharding constraint for transitions
This commit adds code to generate transition sharding constraints
that take into account information from the entire contract, in
particular the global PCM allocation.

This also fixes a bug in the definition of a spurious read. A read
of pseudofield `pf` is only spurious if it flows into a commutative
write to pseudofield `pf`! If the read flows into a write of some
other field, it is not spurious!