Fix Menhir grammar file
Menhir since 20200525 introduced some changes how it treats module aliases.
From the manual:
It is important to note that the header is copied by Menhir only to the .ml file,
not to the .mli file.
Therefore, it should not contain declarations that affect the meaning
of the types that appear in the .mli file.