fix: Resolve the build dependencies of Dockerfile.slim and opam improvements (#1090)
* fix: Resolve the build dependencies of Dockerfile.slim

* opam needs git to work properly

* opam: remove ambiguity in choosing OCaml compiler

The Zilliqa CI shows the following error:

```
(default is 'no', use 'f' to choose a different file) A hook can be added to opam's init scripts to ensure that the shell remains in sync with the opam environment when they are loaded. Set that up? [y/N] y
[ERROR] Compiler selection '4.08.1' is ambiguous. matching packages: { ocaml-base-compiler.4.08.1, ocaml-system.4.08.1 }
make: *** [Makefile:158: opamdep-ci] Error 2
The command '/bin/sh -c bash scripts/in
```