Send contracts and messages via HTTP rather than the file system
Previously, requests would contain a path to a file for:
* Scilla contracts
* Init data JSON
* Message JSON

Instead, we send these directly in the message. The maximum size
of a Scilla contract is ~75KB and the maximum size of init data
and messages is ~135KB.

I've also deleted two utilities that we aren't (yet) using in ZQ2,
so that I don't have to spend time supporting the new request
format for them.