Fix memory leak (#1260)
A long running heap profile of the Scilla process from a ZQ2 node
showed growing memory usage. The allocations were mainly happening
in `add_comment`. I'm not sure why/where we retain the reference
to this forever, but I did see that the only user of `get_comments`
was `scilla-fmt`, which we are not using.

Therefore, I did the easy fix and removed `scilla-fmt`.